requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-anthropic>=0.3.0
langchain-cohere>=0.4.0
langchain-pinecone>=0.2.0
langgraph>=0.3.0
pinecone-client>=5.0.0
cohere>=5.0.0
# Document parsing
pymupdf>=1.24.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
# GCP
google-cloud-firestore>=2.16.0
google-api-python-client>=2.100.0
firebase-admin>=6.0.0
# HTTP
httpx>=0.27.0
# Rate limiting
slowapi>=0.1.9
langchain-experimental
rank-bm25
# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
